Make channel type loading based on class list in settings so deploys can add/remove channel types as they need
